+++ src/model/Shop/ItemList/DailyItemList.java
@@ -6,7 +6,7 @@ import model.Shop.ShopItem;
 public class DailyItemList extends ShopItemList {
     public int remainingTime;
 
-    public DailyItemList(int id) {
+    public DailyItemList(long id) {
         this.id = id;
         //System.out.println(("DailyItemList"));
         this.itemList.add(new ShopItem(ItemDefine.CHESTTYPE, 1, ShopItemDefine.CAN_BUY, ShopItemDefine.DAILY_CHEST_PRICE));
